+++ test/libsolidity/semanticTests/array/pop/byte_array_pop_long_storage_empty_garbage_ref.sol
@@ -19,5 +19,5 @@ contract c {
 // test() ->
 // gas irOptimized: 142640
 // gas legacy: 165363
-// gas legacyOptimized: 159446
+// gas legacyOptimized: 158831
 // storageEmpty -> 1